Use kEmptyObject per upstream suggestion. Move node lookalikes to internals to match node layout.
