+++ internal/errors.js
